+++ Package.swift
@@ -17,7 +17,6 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-nio.git", from: "2.97.1"),
-        .package(url: "https://github.com/apple/swift-nio-ssl.git", from: "2.36.1"),
         .package(url: "https://github.com/apple/swift-crypto.git", from: "4.3.1"),
         .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.36.1"),
         .package(url: "https://github.com/attaswift/BigInt.git", from: "5.5.0"),
@@ -26,17 +25,19 @@ let package = Package(
         .target(
             name: "SwiftATV",
             dependencies: [
-                .product(name: "NIO", package: "swift-nio"),
                 .product(name: "NIOCore", package: "swift-nio"),
                 .product(name: "NIOPosix", package: "swift-nio"),
-                .product(name: "NIOSSL", package: "swift-nio-ssl"),
-                .product(name: "Crypto", package: "swift-crypto"),
-                .product(name: "_CryptoExtras", package: "swift-crypto"),
+                .product(
+                    name: "Crypto",
+                    package: "swift-crypto",
+                    condition: .when(platforms: [.linux])
+                ),
                 .product(name: "SwiftProtobuf", package: "swift-protobuf"),
                 .product(name: "BigInt", package: "BigInt"),
             ],
-            plugins: [
-                .plugin(name: "SwiftProtobufPlugin", package: "swift-protobuf")
+            exclude: [
+                "swift-protobuf-config.json",
+                "Protocols/MRP/Protobuf",
             ]
         ),
         .testTarget(